Moved processing related functions
