+++ pleroma_bot/__init__.py
@@ -2,7 +2,7 @@ import os
 import sys
 import logging
 
-__version__ = "0.7.2"
+__version__ = "0.7.3"
 
 
 class StdOutFilter(logging.Filter):
